backend/Stark.ML/preprocessing/real_time_processor.py
import numpy as np
import pandas as pd
from sklearn.preprocessing import StandardScaler
import tensorflow as tf
from concurrent.futures import ThreadPoolExecutor

class RealTimeProcessor:
    def __init__(self):
        self.scaler = StandardScaler()
        self.feature_extractors = []
        self.batch_size = 32
        self.executor = ThreadPoolExecutor(max_workers=4)
        
    def process_stream(self, data_stream):
        """
        Real-time data processing pipeline
        """
        processed_batches = []
        
        for batch in self._create_batches(data_stream):
            # Process batch asynchronously
            future = self.executor.submit(self._process_batch, batch)
            processed_batches.append(future)
            
            # Yield results as they become available
            for future in processed_batches:
                if future.done():
                    yield future.result()
                    processed_batches.remove(future)
    
    def _process_batch(self, batch):
        # Convert to DataFrame
        df = pd.DataFrame(batch)
        
        # Extract features
        numerical_features = self._process_numerical(df)
        categorical_features = self._process_categorical(df)
        temporal_features = self._extract_temporal_features(df)
        
        # Combine features
        combined_features = np.concatenate([
            numerical_features,
            categorical_features,
            temporal_features
        ], axis=1)
        
        # Scale features
        scaled_features = self.scaler.fit_transform(combined_features)
        
        return {
            'features': scaled_features,
            'metadata': {
                'batch_size': len(batch),
                'timestamp': pd.Timestamp.now(),
                'feature_dims': scaled_features.shape[1]
            }
        }
    
    def _create_batches(self, data_stream):
        batch = []
        for event in data_stream:
            batch.append(event)
            if len(batch) >= self.batch_size:
                yield batch
                batch = []
        if batch:
            yield batch
    
    def _process_numerical(self, df):
        numerical_cols = df.select_dtypes(include=[np.number]).columns
        return df[numerical_cols].fillna(0).values
    
    def _process_categorical(self, df):
        categorical_cols = df.select_dtypes(include=['object']).columns
        return pd.get_dummies(df[categorical_cols]).values
    
    def _extract_temporal_features(self, df):
        if 'timestamp' not in df.columns:
            return np.zeros((len(df), 1))
        
        timestamps = pd.to_datetime(df['timestamp'])
        return np.column_stack([
            timestamps.dt.hour,
            timestamps.dt.dayofweek,
            timestamps.dt.day,
            timestamps.dt.month,
            timestamps.dt.quarter
        ])