requirements.txt
# ML Dependencies
tensorflow>=2.9.1
torch>=1.11.0
transformers>=4.18.0
scikit-learn>=1.0.2
pandas>=1.4.2
numpy>=1.22.3

# Deep Learning
keras>=2.9.0
torchvision>=0.12.0

# Data Processing
scipy>=1.8.1
nltk>=3.7
spacy>=3.3.0

# API and Web
fastapi>=0.78.0
uvicorn>=0.17.6
requests>=2.27.1
aiohttp>=3.8.1

# Monitoring and Logging
prometheus-client>=0.14.1
opentelemetry-api>=1.11.1
opentelemetry-sdk>=1.11.1
opentelemetry-exporter-otlp>=1.11.1
elasticsearch>=7.17.3
kibana>=7.17.3

# Security
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.2
cryptography>=37.0.2

# Database
sqlalchemy>=1.4.36
psycopg2-binary>=2.9.3
alembic>=1.7.7

# Testing
pytest>=7.1.2
pytest-asyncio>=0.18.3
pytest-cov>=3.0.0

# Development
black>=22.3.0
flake8>=4.0.1
mypy>=0.950
isort>=5.10.1

# Documentation
sphinx>=4.5.0
mkdocs>=1.3.0